Change Godot Fill Content to Fit Content Behaviour
Godot by default works as Fill Content, this commit takes license to change that default behaviour to Fit Content.

The problem with Fill Content is that by default container, Labels, button and any type of Control nodes comes by default as fill, this force us to modify the sizing of each control to not fill the content, since filling the area is a contextual behaviour but not desired for all the UI Elements when designing UI.

The Ideal behaviour would be, the button retains its content size by default, the container fits its childs content. If the button expands, the container expands.

GDscriptUI takes the license to work with Fit Content sizing.