Label Fitcontent behaviour and Text Wrapping, stretch_ratio calculation in ContainerBuilder
Label Fitcontent its a tricky implementation and is still not feature complete, so far it works by making the Label expand horizontally at creation, then reading its theme font and calculating the size and the lenght of the text with the function _calculate_stretch_ratio(). This way we can weight how much space the text need in comparison with other labels inside of the container. This weight is applied to size_flags_stretch_ratio to make the bigger text with a higher stretch_ratio, this makes the text ocupy more space   in compasiron with its siblings.

By default all elements inside of a container that has a flag of expand, will have a stretch_ratio of  1.0, by doing this calculation in the Label we are working with a different amount of stretch_ratio.

This approach is still on development and is not a perfect solution. In fact having a Hbox container that has Images, buttons and Labels breaks the expand behaviour because in order to make the elements expand equality they needs to have the same stretch_ratio inside of the container, this means that for a Label that has a ratio of 2.3 the other elements also needs to get to that ratio, this calculation is done inside of BuilderContainer but for now only works with Labels